requirements.txt
Django==1.9.4
django-filter==0.13.0
django-jsonfield==0.9.19
django-model-utils==2.5
djangorestframework==3.3.3
jsonfield==1.0.3
jsonschema==2.5.1
Markdown==2.6.6
psycopg2==2.6.1
requests==2.9.1
